chore: translates sign in form
