+++ app/public/locales/en/a.json
@@ -1,0 +1,7 @@
+{
+  "signUp.form.title": "Passwordless Sign In",
+  "signUp.form.description": "We'll send an email to your inbox with the instructions.",
+  "signUp.footer.slogan": "Banco Satoshi ∞ Bitcoin Investment Banking.",
+  "authEmailForm.emailLabel": "Email",
+  "authEmailForm.button": "Sign In"
+}
