+++ app/public/locales/en/common.json
@@ -5,5 +5,7 @@
   "navbar.businessCampaigns.table": "Table",
   "navbar.howItWorks": "How It Works",
   "navbar.apply": "Apply to Funding",
-  "navbar.logout": "Sign Out"
+  "navbar.logout": "Sign Out",
+  "privacyNotice": "Privacy Notice",
+  "termsAndConditions": "Terms & Conditions"
 }
