+++ app/public/locales/es/a.json
@@ -1,0 +1,7 @@
+{
+  "signUp.form.title": "Inicia Sesión sin Contraseña",
+  "signUp.form.description": "Enviaremos un correo a tu bandeja de entrada con las instrucciones.",
+  "signUp.footer.slogan": "Banco Satoshi ∞ Tu Banco de Inversión y Desarrollo, con Bitcoin.",
+  "authEmailForm.emailLabel": "Correo Electrónico",
+  "authEmailForm.button": "Ingresar"
+}
