+++ app/src/app/auth/sign-up/SignUp.tsx
@@ -11,6 +11,7 @@ import { Button } from "ui/button/Button";
 
 import { SignUpProps } from "./SignUp.types";
 import styles from "./SignUp.module.scss";
+import { useTranslation } from "react-i18next";
 
 const AuthEmailForm = dynamic<AuthEmailFormProps>(
   () => import("ui/form/auth-email-form/AuthEmailForm").then((mod) => mod.AuthEmailForm),
@@ -19,6 +20,7 @@ const AuthEmailForm = dynamic<AuthEmailFormProps>(
 
 export const SignUp: React.FC<SignUpProps> = ({ className }) => {
   const auth = useAuthContext();
+  const { t } = useTranslation(["a", "common"]);
 
   return (
     <div className={clsx(styles["sign-up"], className)}>
@@ -46,10 +48,8 @@ export const SignUp: React.FC<SignUpProps> = ({ className }) => {
               ) : (
                 <Card.Content>
                   <div className={styles["sign-up__instructions"]}>
-                    <Typography.Headline3>Inicia Sesión sin Contraseña</Typography.Headline3>
-                    <Typography.Text>
-                      Enviaremos un correo a tu bandeja de entrada con las instrucciones.
-                    </Typography.Text>
+                    <Typography.Headline3>{t("signUp.form.title")}</Typography.Headline3>
+                    <Typography.Text>{t("signUp.form.description")}</Typography.Text>
                   </div>
                   <AuthEmailForm
                     autoFocus
@@ -60,8 +60,10 @@ export const SignUp: React.FC<SignUpProps> = ({ className }) => {
               )}
             </Card>
             <div className={styles["sign-up__footer"]}>
-              <Typography.Text>Banco Satoshi ∞ Tu Banco de Inversión y Desarrollo, con Bitcoin.</Typography.Text>
-              <Typography.Description>Aviso de Privacidad ∞ Términos y Condiciones</Typography.Description>
+              <Typography.Text>{t("signUp.footer.slogan")}</Typography.Text>
+              <Typography.Description>
+                {t("privacyNotice", { ns: "common" })} ∞ {t("termsAndConditions", { ns: "common" })}
+              </Typography.Description>
             </div>
           </Grid.Col>
         </Grid.Row>
