+++ app/src/app/auth/sign-up/SignUpContainer.tsx
@@ -4,4 +4,4 @@ import { SignUpProps } from "./SignUp.types";
 
 const SignUp = dynamic<SignUpProps>(() => import("./SignUp").then((mod) => mod.SignUp), { ssr: false });
 
-export const SignUpContainer = () => <SignUp>Test</SignUp>;
+export const SignUpContainer = () => <SignUp />;
