+++ app/src/ui/form/auth-email-form/AuthEmailForm.tsx
@@ -5,24 +5,29 @@ import { Button } from "../../button/Button";
 
 import styles from "./AuthEmailForm.module.scss";
 import { AuthEmailFormProps } from "./AuthEmailForm.types";
+import { useTranslation } from "react-i18next";
 
-export const AuthEmailForm: React.FC<AuthEmailFormProps> = ({ className, autoFocus, onSubmit, isLoading }) => (
-  <Form className={clsx(styles["auth-email-form"], className)} onSubmit={onSubmit}>
-    <div className={styles["auth-email-form__inline-wrapper"]}>
-      <div className={styles["auth-email-form__inline-wrapper--form"]}>
-        <Form.TextInput
-          autoFocus={autoFocus}
-          label="Correo Electrónico"
-          id="email"
-          type="email"
-          className={styles["auth-email-form__input"]}
-        />
+export const AuthEmailForm: React.FC<AuthEmailFormProps> = ({ className, autoFocus, onSubmit, isLoading }) => {
+  const { t } = useTranslation("a");
+
+  return (
+    <Form className={clsx(styles["auth-email-form"], className)} onSubmit={onSubmit}>
+      <div className={styles["auth-email-form__inline-wrapper"]}>
+        <div className={styles["auth-email-form__inline-wrapper--form"]}>
+          <Form.TextInput
+            autoFocus={autoFocus}
+            label={t("authEmailForm.emailLabel")}
+            id="email"
+            type="email"
+            className={styles["auth-email-form__input"]}
+          />
+        </div>
+        <div className={styles["auth-email-form__inline-wrapper--button"]}>
+          <Button variant="outlined" fullWidth type="submit" isLoading={isLoading}>
+            {t("authEmailForm.button")}
+          </Button>
+        </div>
       </div>
-      <div className={styles["auth-email-form__inline-wrapper--button"]}>
-        <Button variant="outlined" fullWidth type="submit" isLoading={isLoading}>
-          Ingresar
-        </Button>
-      </div>
-    </div>
-  </Form>
-);
+    </Form>
+  );
+};
